requirements.txt
# BKDict Vocabulary Web Application - Python Dependencies
# Install with: pip install -r requirements.txt

# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# MySQL Database Connector
mysql-connector-python==8.2.0

# Environment Variables Management
python-dotenv==1.0.0

# Additional Utilities
# (lxml is optional - Python's built-in xml.etree.ElementTree is used by default)
# lxml==5.1.0

# Development & Testing (optional)
# pytest==7.4.3
# pytest-flask==1.3.0
